requirements.txt
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.1
boto3==1.22.13
botocore==1.25.13
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.77.1
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
Mako==1.2.0
MarkupSafe==2.1.1
orjson==3.6.8
psycopg2-binary==2.9.3
pydantic==1.9.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
sqlalchemy2-stubs==0.0.2a22
sqlmodel==0.0.6
starlette==0.19.1
typing-extensions==4.2.0
ujson==5.2.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
zipp==3.8.0
